+++ backend/app/Http/Models/Admin.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Http\Models;
 
 use Illuminate\Foundation\Auth\User as Authenticatable;
