+++ backend/config/auth.php
@@ -68,7 +68,7 @@ return [
     'providers' => [
         'users' => [
             'driver' => 'eloquent',
-            'model' => \App\Models\User::class,
+            'model' => \App\Http\Models\User::class,
         ],
 
         // 'users' => [
