+++ backend/database/factories/UserFactory.php
@@ -2,7 +2,7 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Models\User;
+use App\Http\Models\User;
 use Faker\Generator as Faker;
 use Illuminate\Support\Str;
