+++ backend/database/seeds/AdminSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\Admin;
+use App\Http\Models\Admin;
 use Illuminate\Database\Seeder;
 
 class AdminSeeder extends Seeder
